requirement.txt
asgiref==3.4.1
Django==3.1.6
django-filter==21.1
djangorestframework==3.13.1
importlib-metadata==4.8.3
Markdown==3.3.7
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
zipp==3.6.0
